# app/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PtLinker API",
        "description": "PT站点爬虫管理API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/queue/{site_id}/start": {
            "post": {
                "tags": [
                    "queue",
                    "queue"
                ],
                "summary": "启动指定站点的待处理任务",
                "description": "启动队列中指定站点的待处理任务\n\nArgs:\n    site_id: 站点ID\n    db: 数据库会话\n    site_manager: 站点管理器\n    \nReturns:\n    BaseResponse: 包含启动的任务数量",
                "operationId": "start_site_queue_tasks_api_v1_queue__site_id__start_post",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Site Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/queue/clear": {
            "delete": {
                "tags": [
                    "queue",
                    "queue"
                ],
                "summary": "清除待运行的任务队列",
                "description": "清除待运行的任务队列\n\nArgs:\n    site_id: 可选的站点ID，如果不提供则清除所有站点的待运行任务\n    db: 数据库会话\n    site_manager: 站点管理器\n    \nReturns:\n    BaseResponse: 包含清除的任务数量信息",
                "operationId": "clear_pending_tasks_api_v1_queue_clear_delete",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Site Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/credentials/{site_id}": {
            "get": {
                "tags": [
                    "credentials",
                    "credentials"
                ],
                "summary": "获取站点凭证",
                "description": "获取指定站点的凭证信息\n\nArgs:\n    site_id: 站点ID\n    db: 数据库会话\n    \nReturns:\n    CrawlerCredentialResponse: 站点凭证信息",
                "operationId": "get_site_credential_api_v1_credentials__site_id__get",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Site Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CrawlerCredentialResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "credentials",
                    "credentials"
                ],
                "summary": "更新站点凭证",
                "description": "更新指定站点的凭证信息\n\nArgs:\n    site_id: 站点ID\n    credential: 要更新的凭证信息，所有字段都是可选的\n    db: 数据库会话\n    \nReturns:\n    CrawlerCredentialResponse: 更新后的站点凭证信息",
                "operationId": "update_site_credential_api_v1_credentials__site_id__put",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Site Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CrawlerCredentialUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CrawlerCredentialResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BaseResponse": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "success"
                    },
                    "data": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data"
                    },
                    "metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Metadata"
                    }
                },
                "type": "object",
                "title": "BaseResponse"
            },
...
            "CrawlerCredentialResponse": {
                "properties": {
                    "site_id": {
                        "type": "string",
                        "title": "Site Id"
                    },
                    "manual_cookies": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Manual Cookies"
                    },
                    "username": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Username"
                    },
                    "password": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Password"
                    },
                    "authorization": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Authorization"
                    },
                    "apikey": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Apikey"
                    },
                    "enabled": {
                        "type": "boolean",
                        "title": "Enabled"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    }
                },
                "type": "object",
                "required": [
                    "site_id",
                    "enabled"
                ],
                "title": "CrawlerCredentialResponse"
            },
            "CrawlerCredentialUpdate": {
                "properties": {
                    "manual_cookies": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Manual Cookies"
                    },
                    "username": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Username"
                    },
                    "password": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Password"
                    },
                    "authorization": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Authorization"
                    },
                    "apikey": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Apikey"
                    },
                    "enabled": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Enabled"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    }
                },
                "type": "object",
                "title": "CrawlerCredentialUpdate"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}